First set of 3.12. test failures on Mac OS

+++ nimble/core/data/matrixAxis.py
@@ -134,8 +134,8 @@ class MatrixAxis(Axis, metaclass=ABCMeta):
         baseDtype = self._base._data.dtype
         if baseDtype != np.object_ and retDtype == np.object_:
             self._base._data = self._base._data.astype(np.object_)
-        elif baseDtype == np.int_ and retDtype == np.float_:
-            self._base._data = self._base._data.astype(np.float_)
+        elif baseDtype == np.int_ and retDtype == np.float64:
+            self._base._data = self._base._data.astype(np.float64)
         elif baseDtype == np.bool_ and retDtype != np.bool_:
             self._base._data = self._base._data.astype(retDtype)
